+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.1.1 (unreleased)
 
 - Support HTTPS TLS v1.3
+- Support SSH AES256-GCM cipher
 
 ## 0.1.0